Move models and update imports
