+++ examples/react-native/__tests__/ts/Models/Models/Book.ts
