+++ examples/react-native/__tests__/ts/Models/Models/PetOwner.ts
