+++ examples/react-native/__tests__/ts/realm-database/schemas/dictionary-test.tsx
@@ -3,7 +3,7 @@ import {Button, TextInput, View, Text} from 'react-native';
 import {render, fireEvent, waitFor, act} from '@testing-library/react-native';
 import Realm from 'realm';
 import {createRealmContext} from '@realm/react';
-import HomeOwner from '../../models/HomeOwner';
+import HomeOwner from '../../Models/HomeOwner';
 
 const realmConfig = {
   schema: [HomeOwner],
