+++ examples/react-native/__tests__/ts/realm-database/schemas/relationships-test.tsx
@@ -3,8 +3,8 @@ import {View, Text} from 'react-native';
 import {render, waitFor} from '@testing-library/react-native';
 import Realm from 'realm';
 import {createRealmContext} from '@realm/react';
-import User from '../../models/User';
-import Post from '../../models/Post';
+import User from '../../Models/User';
+import Post from '../../Models/Post';
 
 const realmConfig = {
   schema: [User, Post],
